Lazily instantiate some of the axioms via user-propagator

+++ llvm_util/llvm2alive.cpp
@@ -1101,7 +1101,7 @@ public:
         attrs.blockSize = max(attrs.blockSize, asz.getKnownMinSize());
 
         attrs.set(ParamAttrs::Align);
-        attrs.align = max(attrs.align, DL().getABITypeAlignment(ty));
+        attrs.align = max(attrs.align, (uint64_t)DL().getABITypeAlignment(ty));
         continue;
       }
